refactor: nix-files feedback reusing git status, eliminate resulting dead code
