+++ apps/native/src-tauri/src/git/exec.rs
@@ -81,13 +81,6 @@ fn git_command() -> GitCommand {
     GitCommand(cmd)
 }
 
-/// Gets a diff containing only .nix files (including untracked .nix files).
-pub fn get_nix_diff(dir: &str) -> Result<String> {
-    let mut diff = get_tracked_diff(dir, Some("*.nix"))?;
-    append_untracked_diffs(&mut diff, dir, Some("*.nix"))?;
-    Ok(diff)
-}
-
 fn is_safe_repo_relative_path(filename: &str) -> bool {
     let path = Path::new(filename);
     !path.is_absolute()
@@ -111,74 +104,6 @@ pub fn file_diff_contents(dir: &str, filename: &str) -> (String, String) {
         String::new()
     };
     (original, modified)
-}
-
-/// Returns a diff of tracked changes, optionally restricted to a path glob.
-/// Falls back to staged+unstaged when HEAD is unborn (fresh repo with no commits).
-fn get_tracked_diff(dir: &str, path_filter: Option<&str>) -> Result<String> {
-    let extra: Vec<&str> = match path_filter {
-        Some(f) => vec!["--", f],
-        None => vec![],
-    };
-
-    if has_head_commit(dir) {
-        let mut args = vec!["diff", "HEAD"];
-        args.extend(&extra);
-        let output = git_command().args(&args).current_dir(dir).output()?;
-        Ok(String::from_utf8_lossy(&output.stdout).to_string())
-    } else {
-        let mut staged_args = vec!["diff", "--cached"];
-        staged_args.extend(&extra);
-        let mut unstaged_args = vec!["diff"];
-        unstaged_args.extend(&extra);
-
-        let staged = git_command().args(&staged_args).current_dir(dir).output()?;
-        let unstaged = git_command()
-            .args(&unstaged_args)
-            .current_dir(dir)
-            .output()?;
-
-        let mut result = String::from_utf8_lossy(&staged.stdout).to_string();
-        result.push_str(&String::from_utf8_lossy(&unstaged.stdout));
-        Ok(result)
-    }
-}
-
-/// Appends untracked files to `diff` as synthetic new-file diff blocks.
-/// `path_filter` restricts to a glob (e.g. `Some("*.nix")`); `None` includes all files.
-fn append_untracked_diffs(diff: &mut String, dir: &str, path_filter: Option<&str>) -> Result<()> {
-    // --exclude-standard omits gitignored files — hiding them from git status and summarize.
-    let mut args = vec!["ls-files", "--others", "--exclude-standard"];
-    if let Some(filter) = path_filter {
-        args.extend(["--", filter]);
-    }
-
-    let repo_root_dir = repo_root(dir);
-    let untracked_output = git_command()
-        .args(&args)
-        .current_dir(&repo_root_dir)
-        .output()?;
-    let untracked_files = String::from_utf8_lossy(&untracked_output.stdout);
-
-    for file in untracked_files.lines() {
-        if file.is_empty() {
-            continue;
-        }
-        let file_path = repo_root_dir.join(file);
-        if let Ok(contents) = std::fs::read_to_string(&file_path) {
-            diff.push_str(&format!("\ndiff --git a/{} b/{}\n", file, file));
-            diff.push_str("new file mode 100644\n");
-            diff.push_str("--- /dev/null\n");
-            diff.push_str(&format!("+++ b/{}\n", file));
-            let line_count = contents.lines().count();
-            diff.push_str(&format!("@@ -0,0 +1,{} @@\n", line_count));
-            for line in contents.lines() {
-                diff.push_str(&format!("+{}\n", line));
-            }
-        }
-    }
-
-    Ok(())
 }
 
 /// Registers all untracked files as intent-to-add in the git index.
@@ -527,48 +452,6 @@ mod tests {
 
         let (_, modified) = file_diff_contents(&repo_dir_str, "flake.nix");
         assert_eq!(modified, "{ inputs = {}; }");
-    }
-
-    #[test]
-    fn test_get_nix_diff_excludes_non_nix_untracked() {
-        let temp_dir = TempDir::new().unwrap();
-        let repo_dir = temp_dir.path().join("repo");
-        let repo_dir_str = repo_dir.to_string_lossy().to_string();
-        init_repo(&repo_dir_str).unwrap();
-
-        fs::write(repo_dir.join("flake.nix"), "{ }").unwrap();
-        commit_all(&repo_dir_str, "initial").unwrap();
-        fs::write(repo_dir.join("config.nix"), "{ }").unwrap();
-        fs::write(repo_dir.join("readme.txt"), "hello").unwrap();
-
-        let diff = get_nix_diff(&repo_dir_str).unwrap();
-        assert!(
-            diff.contains("config.nix"),
-            ".nix untracked file should appear"
-        );
-        assert!(
-            !diff.contains("readme.txt"),
-            "non-.nix file should be excluded"
-        );
-    }
-
-    #[test]
-    fn test_get_nix_diff_excludes_gitignored_nix_file() {
-        let temp_dir = TempDir::new().unwrap();
-        let repo_dir = temp_dir.path().join("repo");
-        let repo_dir_str = repo_dir.to_string_lossy().to_string();
-        init_repo(&repo_dir_str).unwrap();
-
-        fs::write(repo_dir.join(".gitignore"), "secret.nix\n").unwrap();
-        fs::write(repo_dir.join("flake.nix"), "{ }").unwrap();
-        commit_all(&repo_dir_str, "initial").unwrap();
-        fs::write(repo_dir.join("secret.nix"), "{ password = \"123\"; }").unwrap();
-
-        let diff = get_nix_diff(&repo_dir_str).unwrap();
-        assert!(
-            !diff.contains("secret.nix"),
-            "gitignored .nix file must not appear in diff"
-        );
     }
 
     #[test]
